+++ lib/scripts/components/screens/screensList/game/style/_animation.scss
@@ -1,7 +1,7 @@
 //  animation flip-in-ver-left
 
 .flip-in-ver-left {
-    -webkit-animation: flip-in-ver-left 0.5s cubic-bezier(0.250, 0.460, 0.450, 0.940) both;
+    // -webkit-animation: flip-in-ver-left 0.5s cubic-bezier(0.250, 0.460, 0.450, 0.940) both;
     animation: flip-in-ver-left 0.5s cubic-bezier(0.250, 0.460, 0.450, 0.940) both;
 }
