+++ lib/scripts/components/screens/screensList/game/style/gameStyle.scss
@@ -1,0 +1,113 @@
+// ----------------------- GENERAL FOR BODY ----------------------
+//
+
+@import 'animation';
+
+
+$color-bg: #004980;
+
+$color-block-1: #c2f5ff;
+$color-block-2: #fff;
+
+$color-text-1: $color-bg;
+$color-text-2: #0080c1;
+$color-text-3: $color-block-2;
+$color-text-4: #000;
+
+$color-all-1: #7ac100;
+
+
+
+.screen_game {
+
+  &_header {
+    width: 100%;
+    height: max-content;
+    min-width: 760px;
+
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+
+    padding-left: 100px;
+    padding-right: 100px;
+
+
+  }
+
+  &_gameField {
+    width: 100%;
+    height: 640px;
+
+    border: 0px solid black;
+
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+
+    padding: 10px;
+
+    overflow: visible;
+
+    &_cardsSetBox {
+      &_1 {
+        display: grid;
+
+        grid-template-columns: repeat(3, 140px);
+        grid-template-rows: repeat(2, 200px);
+
+        gap: 15px;
+
+        overflow: visible;
+      }
+
+      &_2 {
+        display: grid;
+
+        grid-template-columns: repeat(4, 120px);
+        grid-template-rows: repeat(3, 165px);
+
+        gap: 15px;
+
+        overflow: visible;
+      }
+
+      &_3 {
+        display: grid;
+
+        grid-template-columns: repeat(6, 120px);
+        grid-template-rows: repeat(3, 165px);
+
+        gap: 15px;
+
+        overflow: visible;
+      }
+
+      &_4 {
+        display: grid;
+
+        grid-template-columns: repeat(9, 100px);
+        grid-template-rows: repeat(4, 140px);
+
+        gap: 15px;
+
+        overflow: visible;
+      }
+    }
+
+    &_cardBox {
+      cursor: pointer;
+
+      & img {
+        width: 100%;
+        height: 100%;
+      }
+
+      &:hover {
+        transform: scale(1.1);
+      }
+    }
+  }
+}
